refactor: merge TaskManager into TaskService
