+++ src/main/java/uni/usic/application/service/TaskService.java
@@ -3,13 +3,21 @@ package uni.usic.application.service;
 import uni.usic.domain.entity.maintasks.Task;
 import uni.usic.domain.enums.TaskPriority;
 import uni.usic.domain.enums.TaskProgress;
+import uni.usic.infrastructure.repository.TaskFileRepository;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 public class TaskService implements TaskOperations {
+    //    private static List<Task> taskList = new ArrayList<>();
+    private static Map<String, Task> taskMap = new HashMap<>();
+
     @Override
     public void viewTask(String id) {
-        Task task = TaskManager.getTaskById(id);
+        Task task = getTaskById(id);
         if(task == null) {
             System.out.println("No task has found like task ID: " + id);
             return;
@@ -20,12 +28,23 @@ public class TaskService implements TaskOperations {
     @Override
     public Task createTask(String title, String description, LocalDate startDate, LocalDate endDate, TaskPriority priority) {
         Task task = new Task(title, description, startDate, endDate, priority);
-        return task;
+        boolean result = addTask(task);
+
+        if(result) {
+            return task;
+        } else {
+            return null;
+        }
+    }
+
+    public boolean addTask(Task task) {
+        TaskFileRepository taskFileRepository = new TaskFileRepository();
+        return taskFileRepository.save(task);
     }
 
     @Override
     public boolean modifyTask(String id, String title, String description, LocalDate startDate, LocalDate endDate, TaskPriority priority, TaskProgress progress, Integer reminderDaysBefore) {
-        Task task = TaskManager.getTaskById(id);
+        Task task = getTaskById(id);
         if(task == null) {
             return false;
         }
@@ -41,7 +60,7 @@ public class TaskService implements TaskOperations {
 
     @Override
     public boolean updateProgress(String id, TaskProgress progress) {
-        Task task = TaskManager.getTaskById(id);
+        Task task = getTaskById(id);
         if(task == null) {
             return false;
         }
@@ -51,6 +70,25 @@ public class TaskService implements TaskOperations {
 
     @Override
     public boolean deleteTask(String id) {
-        return TaskManager.removeTask(id);
+        return removeTask(id);
+    }
+
+    public static List<Task> viewTaskList() {
+        return new ArrayList<>(taskMap.values());
+    }
+
+    // TODO: getTaskList()?
+    public static List<Task> convertTaskMapToList(Map<String, Task> taskMap) {
+        return new ArrayList<>(taskMap.values()); // Convert map values to list
+    }
+
+    public static Task getTaskById(String taskId) {
+        return taskMap.get(taskId);
+    }
+
+
+
+    public static boolean removeTask(String id) {
+        return taskMap.remove(id) != null;
     }
 }
